requirements.txt
pillow
pandas
pytorch_model_summary
albumentations
tqdm
numpy
torch
torchvision
timm
segmentation-models-pytorch
wandb
matplotlib
